+++ app/client-app/src/app/routes/root.jsx
@@ -7,7 +7,7 @@ const Root = () => {
   if (user === undefined) {
     return <BusyPage verb="Checking" noun="user" />;
   } else {
-    if (isAdmin) navigate("/admin/");
+    if (isAdmin) navigate("/admin/surveys");
     else navigate("/survey");
     return null;
   }
